requirements.txt
mistralai>=1.0.0
chromadb>=0.4.0
streamlit>=1.30.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
